+++ src/Symfony/Component/Process/ProcessBuilder.php
@@ -24,7 +24,7 @@ class ProcessBuilder
     private $arguments;
     private $cwd;
     private $env = array();
-    private $stdin;
+    private $input;
     private $timeout = 60;
     private $options = array();
     private $inheritEnv = true;
@@ -156,13 +156,13 @@ class ProcessBuilder
     /**
      * Sets the input of the process.
      *
-     * @param string $stdin The input as a string
+     * @param string $input The input as a string
      *
      * @return ProcessBuilder
      */
-    public function setInput($stdin)
+    public function setInput($input)
     {
-        $this->stdin = $stdin;
+        $this->input = $input;
 
         return $this;
     }
@@ -261,7 +261,7 @@ class ProcessBuilder
             $env = $this->env;
         }
 
-        $process = new Process($script, $this->cwd, $env, $this->stdin, $this->timeout, $options);
+        $process = new Process($script, $this->cwd, $env, $this->input, $this->timeout, $options);
 
         if ($this->outputDisabled) {
             $process->disableOutput();